feat: integrated PC builds with shared preferences
This commit introduces several key changes to integrate PC builds with shared preferences.

- **Build Data Persistence:**
  - The `BuzyUserBuildPrefManager` class is created to manage saving and retrieving PC builds for users using shared preferences.
  - It uses Gson for serialization and deserialization of PCBuild objects.
  - A custom type adapter is implemented to handle ClosedFloatingPointRange when serializing PCBuild data.

- **Build List Management:**
  - Functions to load the build list are added to the `UserDetailUtil` class.
  - The `TrackerFragment` and `BuildListFragment` are updated to use the shared preferences for retrieving build lists instead of an in-memory list managed by the user manager.
  - The `BuildTrackerRecyclerViewAdapter` is updated to correctly bind build data.

- **Build Creation and Saving:**
  - In `NewBuildFragment`, the build budget is now saved as a Double to prevent errors.
  - In `NewBuildSummaryActivity`, build data is now saved using `BuzyUserBuildPrefManager` when the save button is clicked.
  - Also implemented is the generation of unique build IDs using a utility function.

- **Build Loading and Display:**
  - The `BuildListFragment` now retrieves and displays the list of builds.
  - The build summary activity now displays information about the selected build.

- **Session Management:**
  - The `SessionManager` class now returns a non-nullable string for the username.

- **Settings and Logout:**
  - The `SettingsFragment` is updated to remove handling of the last logged-in user and theme preference directly. This logic is probably moved to LogoutPromptActivity in another commit (follow-up).

- **App Session:**
  - The `BuzyUserAppSession` is updated to store the selected build and build list.

+++ app/src/main/java/io/buzypc/app/UI/Navigation/Fragments/NewBuild/NewBuildFragment.kt
@@ -52,7 +52,7 @@ class NewBuildFragment : Fragment() {
             }
             requireContext()
             (context?.applicationContext as BuzyUserAppSession).buildName = etBuildName.text.toString()
-            (context?.applicationContext as BuzyUserAppSession).buildBudget = etBudgetInput.text.toString()
+            (context?.applicationContext as BuzyUserAppSession).buildBudget = etBudgetInput.text.toString().toDouble()
             val intent = Intent(requireActivity(), NewBuildSummaryActivity::class.java)
             startActivity(intent)
         }
